# swagger/swagger20.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "gopaytest code test payments API",
    "version" : "1.0.0",
    "title" : "Payments API",
    "termsOfService" : "",
    "contact" : {
      "name" : "Francesco Casula",
      "email" : "fra.casula@gmail.com"
    }
  },
  "host" : "api.test.gopaytest.tech",
  "basePath" : "/v1",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
    "/payments/" : {
      "get" : {
        "tags" : [ "Payments" ],
        "summary" : "Get the list of Payments",
        "description" : "Loads a list of Payments",
        "parameters" : [ {
          "name" : "$size",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "description" : "Size of the page to retrieve.",
          "x-example" : 10
        }, {
          "name" : "$page",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "description" : "Number of the page to retrieve.",
          "x-example" : 1
        } ],
        "responses" : {
          "200" : {
            "description" : "Status 200",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Payment"
              }
            },
            "examples" : {
              "application/json" : "{\n    \"data\": [\n        {\n            \"type\": \"Payment\",\n            \"id\": \"4ee3a8d8-ca7b-4290-a52c-dd5b6165ec43\",\n            \"version\": 0,\n            \"organisation_id\": \"743d5b63-8e6f-432e-a8fa-c5d8d2ee5fcb\",\n            \"attributes\": {\n                \"amount\": \"100.21\",\n                \"beneficiary_party\": {\n                    \"account_name\": \"W Owens\",\n                    \"account_number\": \"31926819\",\n                    \"account_number_code\": \"BBAN\",\n                    \"account_type\": 0,\n                    \"address\": \"1 The Beneficiary Localtown SE2\",\n                    \"bank_id\": \"403000\",\n                    \"bank_id_code\": \"GBDSC\",\n                    \"name\": \"Wilfred Jeremiah Owens\"\n                },\n                \"charges_information\": {\n                    \"bearer_code\": \"SHAR\",\n                    \"sender_charges\": [\n                        {\n                            \"amount\": \"5.00\",\n                            \"currency\": \"GBP\"\n                        },\n                        {\n                            \"amount\": \"10.00\",\n                            \"currency\": \"USD\"\n                        }\n                    ],\n                    \"receiver_charges_amount\": \"1.00\",\n                    \"receiver_charges_currency\": \"USD\"\n                },\n                \"currency\": \"GBP\",\n                \"debtor_party\": {\n                    \"account_name\": \"EJ Brown Black\",\n                    \"account_number\": \"GB29XABC10161234567801\",\n                    \"account_number_code\": \"IBAN\",\n                    \"address\": \"10 Debtor Crescent Sourcetown NE1\",\n                    \"bank_id\": \"203301\",\n                    \"bank_id_code\": \"GBDSC\",\n                    \"name\": \"Emelia Jane Brown\"\n                },\n                \"end_to_end_reference\": \"Wil piano Jan\",\n                \"fx\": {\n                    \"contract_reference\": \"FX123\",\n                    \"exchange_rate\": \"2.00000\",\n                    \"original_amount\": \"200.42\",\n                    \"original_currency\": \"USD\"\n                },\n                \"numeric_reference\": \"1002001\",\n                \"payment_id\": \"123456789012345678\",\n                \"payment_purpose\": \"Paying for goods/services\",\n                \"payment_scheme\": \"FPS\",\n                \"payment_type\": \"Credit\",\n                \"processing_date\": \"2017-01-18\",\n                \"reference\": \"Payment for Em's piano lessons\",\n                \"scheme_payment_sub_type\": \"InternetBanking\",\n                \"scheme_payment_type\": \"ImmediatePayment\",\n                \"sponsor_party\": {\n                    \"account_number\": \"56781234\",\n                    \"bank_id\": \"123123\",\n                    \"bank_id_code\": \"GBDSC\"\n                }\n            }\n        }\n    ],\n    \"links\": {\n        \"self\": \"https://api.test.gopaytest.tech/v1/payments\"\n    }\n}\n"
            },
            "headers" : {
              "X-Page-Count" : {
                "type" : "integer",
                "x-example" : 1
              },
              "X-Page-Number" : {
                "type" : "integer",
                "x-example" : 1
              },
              "X-Page-Size" : {
                "type" : "integer",
                "x-example" : 25
              },
              "X-Total-Count" : {
                "type" : "integer",
                "x-example" : 2
              }
            }
          },
          "500" : {
            "description" : "Status 500",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Payments" ],
        "summary" : "Create a Payment",
        "description" : "Adds a Payment",
        "consumes" : [ ],
        "parameters" : [ {
          "name" : "body",
          "in" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Payment"
          },
          "x-examples" : {
            "application/json" : "{\n    \"type\": \"Payment\",\n    \"organisation_id\": \"743d5b63-8e6f-432e-a8fa-c5d8d2ee5fcb\",\n    \"attributes\": {\n        \"amount\": \"100.21\",\n        \"beneficiary_party\": {\n            \"account_name\": \"W Owens\",\n            \"account_number\": \"31926819\",\n            \"account_number_code\": \"BBAN\",\n            \"account_type\": 0,\n            \"address\": \"1 The Beneficiary Localtown SE2\",\n            \"bank_id\": \"403000\",\n            \"bank_id_code\": \"GBDSC\",\n            \"name\": \"Wilfred Jeremiah Owens\"\n        },\n        \"charges_information\": {\n            \"bearer_code\": \"SHAR\",\n            \"sender_charges\": [\n                {\n                    \"amount\": \"5.00\",\n                    \"currency\": \"GBP\"\n                },\n                {\n                    \"amount\": \"10.00\",\n                    \"currency\": \"USD\"\n                }\n            ],\n            \"receiver_charges_amount\": \"1.00\",\n            \"receiver_charges_currency\": \"USD\"\n        },\n        \"currency\": \"GBP\",\n        \"debtor_party\": {\n            \"account_name\": \"EJ Brown Black\",\n            \"account_number\": \"GB29XABC10161234567801\",\n            \"account_number_code\": \"IBAN\",\n            \"address\": \"10 Debtor Crescent Sourcetown NE1\",\n            \"bank_id\": \"203301\",\n            \"bank_id_code\": \"GBDSC\",\n            \"name\": \"Emelia Jane Brown\"\n        },\n        \"end_to_end_reference\": \"Wil piano Jan\",\n        \"fx\": {\n            \"contract_reference\": \"FX123\",\n            \"exchange_rate\": \"2.00000\",\n            \"original_amount\": \"200.42\",\n            \"original_currency\": \"USD\"\n        },\n        \"numeric_reference\": \"1002001\",\n        \"payment_id\": \"123456789012345678\",\n        \"payment_purpose\": \"Paying for goods/services\",\n        \"payment_scheme\": \"FPS\",\n        \"payment_type\": \"Credit\",\n        \"processing_date\": \"2017-01-18\",\n        \"reference\": \"Payment for Em's piano lessons\",\n        \"scheme_payment_sub_type\": \"InternetBanking\",\n        \"scheme_payment_type\": \"ImmediatePayment\",\n        \"sponsor_party\": {\n            \"account_number\": \"56781234\",\n            \"bank_id\": \"123123\",\n            \"bank_id_code\": \"GBDSC\"\n        }\n    }\n}"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Status 201",
            "schema" : {
              "$ref" : "#/definitions/Payment"
            },
            "examples" : {
              "application/json" : "{\n    \"data\": {\n        \"type\": \"Payment\",\n        \"id\": \"cca3b50b-4acb-43dd-8465-8e7d6eac2cac\",\n        \"version\": 1,\n        \"organisation_id\": \"743d5b63-8e6f-432e-a8fa-c5d8d2ee5fcb\",\n        \"attributes\": {\n            \"amount\": \"100.21\",\n            \"beneficiary_party\": {\n                \"account_name\": \"W Owens\",\n                \"account_number\": \"31926819\",\n                \"account_number_code\": \"BBAN\",\n                \"account_type\": 0,\n                \"address\": \"1 The Beneficiary Localtown SE2\",\n                \"bank_id\": \"403000\",\n                \"bank_id_code\": \"GBDSC\",\n                \"name\": \"Wilfred Jeremiah Owens\"\n            },\n            \"charges_information\": {\n                \"bearer_code\": \"SHAR\",\n                \"sender_charges\": [\n                    {\n                        \"amount\": \"5.00\",\n                        \"currency\": \"GBP\"\n                    },\n                    {\n                        \"amount\": \"10.00\",\n                        \"currency\": \"USD\"\n                    }\n                ],\n                \"receiver_charges_amount\": \"1.00\",\n                \"receiver_charges_currency\": \"USD\"\n            },\n            \"currency\": \"GBP\",\n            \"debtor_party\": {\n                \"account_name\": \"EJ Brown Black\",\n                \"account_number\": \"GB29XABC10161234567801\",\n                \"account_number_code\": \"IBAN\",\n                \"address\": \"10 Debtor Crescent Sourcetown NE1\",\n                \"bank_id\": \"203301\",\n                \"bank_id_code\": \"GBDSC\",\n                \"name\": \"Emelia Jane Brown\"\n            },\n            \"end_to_end_reference\": \"Wil piano Jan\",\n            \"fx\": {\n                \"contract_reference\": \"FX123\",\n                \"exchange_rate\": \"2.00000\",\n                \"original_amount\": \"200.42\",\n                \"original_currency\": \"USD\"\n            },\n            \"numeric_reference\": \"1002001\",\n            \"payment_id\": \"123456789012345678\",\n            \"payment_purpose\": \"Paying for goods/services\",\n            \"payment_scheme\": \"FPS\",\n            \"payment_type\": \"Credit\",\n            \"processing_date\": \"2017-01-18\",\n            \"reference\": \"Payment for Em's piano lessons\",\n            \"scheme_payment_sub_type\": \"InternetBanking\",\n            \"scheme_payment_type\": \"ImmediatePayment\",\n            \"sponsor_party\": {\n                \"account_number\": \"56781234\",\n                \"bank_id\": \"123123\",\n                \"bank_id_code\": \"GBDSC\"\n            }\n        }\n    },\n    \"links\": {\n        \"self\": \"https://api.test.gopaytest.tech/v1/payments/cca3b50b-4acb-43dd-8465-8e7d6eac2cac\"\n    }\n}"
            },
            "headers" : {
              "Location" : {
                "type" : "string",
                "x-example" : "https://api.test.gopaytest.tech/v1/payments/4ee3a8d8-ca7b-4290-a52c-dd5b6165ec43"
              }
            }
          },
          "400" : {
            "description" : "Status 400",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Status 500",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/payments/{paymentId}" : {
      "get" : {
        "tags" : [ "Payments" ],
        "summary" : "Load an individual Payment",
        "description" : "Loads a Payment",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Status 200",
            "schema" : {
              "$ref" : "#/definitions/Payment"
            },
            "examples" : {
              "application/json" : "{\n    \"data\": {\n        \"type\": \"Payment\",\n        \"id\": \"ee761e4b-2b5a-411c-84c7-1ac1d48854fe\",\n        \"version\": 3,\n        \"organisation_id\": \"743d5b63-8e6f-432e-a8fa-c5d8d2ee5fcb\",\n        \"attributes\": {\n            \"amount\": \"100.21\",\n            \"beneficiary_party\": {\n                \"account_name\": \"W Owens\",\n                \"account_number\": \"31926819\",\n                \"account_number_code\": \"BBAN\",\n                \"account_type\": 0,\n                \"address\": \"1 The Beneficiary Localtown SE2\",\n                \"bank_id\": \"403000\",\n                \"bank_id_code\": \"GBDSC\",\n                \"name\": \"Wilfred Jeremiah Owens\"\n            },\n            \"charges_information\": {\n                \"bearer_code\": \"SHAR\",\n                \"sender_charges\": [\n                    {\n                        \"amount\": \"5.00\",\n                        \"currency\": \"GBP\"\n                    },\n                    {\n                        \"amount\": \"10.00\",\n                        \"currency\": \"USD\"\n                    }\n                ],\n                \"receiver_charges_amount\": \"1.00\",\n                \"receiver_charges_currency\": \"USD\"\n            },\n            \"currency\": \"GBP\",\n            \"debtor_party\": {\n                \"account_name\": \"EJ Brown Black\",\n                \"account_number\": \"GB29XABC10161234567801\",\n                \"account_number_code\": \"IBAN\",\n                \"address\": \"10 Debtor Crescent Sourcetown NE1\",\n                \"bank_id\": \"203301\",\n                \"bank_id_code\": \"GBDSC\",\n                \"name\": \"Emelia Jane Brown\"\n            },\n            \"end_to_end_reference\": \"Wil piano Jan\",\n            \"fx\": {\n                \"contract_reference\": \"FX123\",\n                \"exchange_rate\": \"2.00000\",\n                \"original_amount\": \"200.42\",\n                \"original_currency\": \"USD\"\n            },\n            \"numeric_reference\": \"1002001\",\n            \"payment_id\": \"123456789012345678\",\n            \"payment_purpose\": \"Paying for goods/services\",\n            \"payment_scheme\": \"FPS\",\n            \"payment_type\": \"Credit\",\n            \"processing_date\": \"2017-01-18\",\n            \"reference\": \"Payment for Em's piano lessons\",\n            \"scheme_payment_sub_type\": \"InternetBanking\",\n            \"scheme_payment_type\": \"ImmediatePayment\",\n            \"sponsor_party\": {\n                \"account_number\": \"56781234\",\n                \"bank_id\": \"123123\",\n                \"bank_id_code\": \"GBDSC\"\n            }\n        }\n    },\n    \"links\": {\n        \"self\": \"https://api.test.gopaytest.tech/v1/payments/ee761e4b-2b5a-411c-84c7-1ac1d48854fe\"\n    }\n}"
            }
          },
          "400" : {
            "description" : "Status 400",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Status 404",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Status 500",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Payments" ],
        "summary" : "Update a Payment",
        "description" : "Stores a Payment",
        "consumes" : [ ],
        "parameters" : [ {
          "name" : "body",
          "in" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Payment"
          },
          "x-examples" : {
            "application/json" : "{\n    \"type\": \"Payment\",\n    \"organisation_id\": \"743d5b63-8e6f-432e-a8fa-c5d8d2ee5fcb\",\n    \"attributes\": {\n        \"amount\": \"100.21\",\n        \"beneficiary_party\": {\n            \"account_name\": \"W Owens\",\n            \"account_number\": \"31926819\",\n            \"account_number_code\": \"BBAN\",\n            \"account_type\": 0,\n            \"address\": \"1 The Beneficiary Localtown SE2\",\n            \"bank_id\": \"403000\",\n            \"bank_id_code\": \"GBDSC\",\n            \"name\": \"Wilfred Jeremiah Owens\"\n        },\n        \"charges_information\": {\n            \"bearer_code\": \"SHAR\",\n            \"sender_charges\": [\n                {\n                    \"amount\": \"5.00\",\n                    \"currency\": \"GBP\"\n                },\n                {\n                    \"amount\": \"10.00\",\n                    \"currency\": \"USD\"\n                }\n            ],\n            \"receiver_charges_amount\": \"1.00\",\n            \"receiver_charges_currency\": \"USD\"\n        },\n        \"currency\": \"GBP\",\n        \"debtor_party\": {\n            \"account_name\": \"EJ Brown Black\",\n            \"account_number\": \"GB29XABC10161234567801\",\n            \"account_number_code\": \"IBAN\",\n            \"address\": \"10 Debtor Crescent Sourcetown NE1\",\n            \"bank_id\": \"203301\",\n            \"bank_id_code\": \"GBDSC\",\n            \"name\": \"Emelia Jane Brown\"\n        },\n        \"end_to_end_reference\": \"Wil piano Jan\",\n        \"fx\": {\n            \"contract_reference\": \"FX123\",\n            \"exchange_rate\": \"2.00000\",\n            \"original_amount\": \"200.42\",\n            \"original_currency\": \"USD\"\n        },\n        \"numeric_reference\": \"1002001\",\n        \"payment_id\": \"123456789012345678\",\n        \"payment_purpose\": \"Paying for goods/services\",\n        \"payment_scheme\": \"FPS\",\n        \"payment_type\": \"Credit\",\n        \"processing_date\": \"2017-01-18\",\n        \"reference\": \"Payment for Em's piano lessons\",\n        \"scheme_payment_sub_type\": \"InternetBanking\",\n        \"scheme_payment_type\": \"ImmediatePayment\",\n        \"sponsor_party\": {\n            \"account_number\": \"56781234\",\n            \"bank_id\": \"123123\",\n            \"bank_id_code\": \"GBDSC\"\n        }\n    }\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Status 200",
            "schema" : {
              "$ref" : "#/definitions/Payment"
            },
            "examples" : {
              "application/json" : "{\n    \"data\": {\n        \"type\": \"Payment\",\n        \"id\": \"cca3b50b-4acb-43dd-8465-8e7d6eac2cac\",\n        \"version\": 2,\n        \"organisation_id\": \"743d5b63-8e6f-432e-a8fa-c5d8d2ee5fcb\",\n        \"attributes\": {\n            \"amount\": \"100.21\",\n            \"beneficiary_party\": {\n                \"account_name\": \"W Owens\",\n                \"account_number\": \"31926819\",\n                \"account_number_code\": \"BBAN\",\n                \"account_type\": 0,\n                \"address\": \"1 The Beneficiary Localtown SE2\",\n                \"bank_id\": \"403000\",\n                \"bank_id_code\": \"GBDSC\",\n                \"name\": \"Wilfred Jeremiah Owens\"\n            },\n            \"charges_information\": {\n                \"bearer_code\": \"SHAR\",\n                \"sender_charges\": [\n                    {\n                        \"amount\": \"5.00\",\n                        \"currency\": \"GBP\"\n                    },\n                    {\n                        \"amount\": \"10.00\",\n                        \"currency\": \"USD\"\n                    }\n                ],\n                \"receiver_charges_amount\": \"1.00\",\n                \"receiver_charges_currency\": \"USD\"\n            },\n            \"currency\": \"GBP\",\n            \"debtor_party\": {\n                \"account_name\": \"EJ Brown Black\",\n                \"account_number\": \"GB29XABC10161234567801\",\n                \"account_number_code\": \"IBAN\",\n                \"address\": \"10 Debtor Crescent Sourcetown NE1\",\n                \"bank_id\": \"203301\",\n                \"bank_id_code\": \"GBDSC\",\n                \"name\": \"Emelia Jane Brown\"\n            },\n            \"end_to_end_reference\": \"Wil piano Jan\",\n            \"fx\": {\n                \"contract_reference\": \"FX123\",\n                \"exchange_rate\": \"2.00000\",\n                \"original_amount\": \"200.42\",\n                \"original_currency\": \"USD\"\n            },\n            \"numeric_reference\": \"1002001\",\n            \"payment_id\": \"123456789012345678\",\n            \"payment_purpose\": \"Paying for goods/services\",\n            \"payment_scheme\": \"FPS\",\n            \"payment_type\": \"Credit\",\n            \"processing_date\": \"2017-01-18\",\n            \"reference\": \"Payment for Em's piano lessons\",\n            \"scheme_payment_sub_type\": \"InternetBanking\",\n            \"scheme_payment_type\": \"ImmediatePayment\",\n            \"sponsor_party\": {\n                \"account_number\": \"56781234\",\n                \"bank_id\": \"123123\",\n                \"bank_id_code\": \"GBDSC\"\n            }\n        }\n    },\n    \"links\": {\n        \"self\": \"https://api.test.gopaytest.tech/v1/payments/cca3b50b-4acb-43dd-8465-8e7d6eac2cac\"\n    }\n}"
            }
          },
          "400" : {
            "description" : "Status 400",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Status 404",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "409" : {
            "description" : "An update has been issued with stale data, the client should fetch the updated resource and apply the changes again.",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Status 500",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Payments" ],
        "summary" : "Delete a Payment",
        "description" : "Deletes a Payment",
        "parameters" : [ ],
        "responses" : {
          "204" : {
            "description" : "Status 204"
          },
          "400" : {
            "description" : "Status 400",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Status 404",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Status 500",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "parameters" : [ {
        "name" : "paymentId",
        "in" : "path",
        "required" : true,
        "type" : "string",
        "description" : "Identifier of the Payment",
        "x-example" : "0e8dd830-ad98-11e6-bf2e-47644ada7c0f"
      } ]
    }
  },
  "securityDefinitions" : {
    "None" : {
      "description" : "This API is not secured, it's not meant to be exposed outside of the cluster. Authentication should be handled via another API (e.g. API Gateway)."
    }
  },
  "definitions" : {
    "Error" : {
      "type" : "object",
      "required" : [ "code", "reasonPhrase" ],
      "properties" : {
        "code" : {
          "type" : "integer",
          "minimum" : 400,
          "maximum" : 599
        },
        "description" : {
          "type" : "string",
          "example" : "Bad query parameter [$size]: Invalid integer value [abc]"
        },
        "reasonPhrase" : {
          "type" : "string",
          "example" : "Bad Request"
        }
      },
      "description" : "This general error structure is used throughout this API.",
      "example" : "{\n  \"code\": 400,\n  \"description\": \"Bad query parameter [$size]: Invalid integer value [abc]\",\n  \"reasonPhrase\": \"Bad Request\"\n}"
    },
...
  }
}